odyssey_video/pyproject.toml
[project]
name = "odyssey-video"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=9.0",
    "numpy>=1.21",
    "imageio>=2.16",
    "imageio-ffmpeg>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
